requirements.txt
Pillow==10.2.0
torch==2.2.0
torchvision==0.17.0
